Refactor LIS3DH Accelerometer blocks to new module.
