+++ src/modules/editor.js
@@ -50,6 +50,7 @@ import './blockly/generators/propc/sensors/fingerprint_scanner';
 import './blockly/generators/propc/sensors/gps';
 import './blockly/generators/propc/sensors/hmc5883l_compass';
 import './blockly/generators/propc/sensors/joystick';
+import './blockly/generators/propc/sensors/lis3dh_accelerometer';
 import './blockly/generators/propc/sensors/mma7455';
 import './blockly/generators/propc/sensors/mx2125';
 import './blockly/generators/propc/sensors/ping';
@@ -1433,7 +1434,6 @@ export function createNewProject() {
   // Save the form fields into the projectData object
   // The projectData variable is defined in globals.js
   const projectName = $('#new-project-name').val();
-  const createdDateHtml = $('#edit-project-created-date').html();
   const description = $('#new-project-description').val();
   const boardType = $('#new-project-board-type').val();
 
@@ -1451,8 +1451,8 @@ export function createNewProject() {
         tmpBoardType,
         ProjectTypes.PROPC,
         '',
-        createdDateHtml,
-        createdDateHtml,
+        date,
+        date,
         timestamp,
         true);
